+++ src/Wheel.java
@@ -1,17 +1,17 @@
 
 public class Wheel {
 
-	short rimDiameter;
+	int rimDiameter;
 	
-	short tireWidth;
+	int tireWidth;
 	
-	short tireDiameter;
+	int tireDiameter;
 	
 	String modelNumber;
 	
 	double price;
 
-	public Wheel(short rimDiameter, short tireWidth, short tireDiameter, String modelNumber, double price) {
+	public Wheel(int rimDiameter, int tireWidth, int tireDiameter, String modelNumber, double price) {
 		this.rimDiameter = rimDiameter;
 		this.tireWidth = tireWidth;
 		this.tireDiameter = tireDiameter;
